requirements.txt
fastapi==0.112.1
json_repair==0.25.3
ollama==0.3.1
openai==1.42.0
peewee==3.17.6
pydantic==2.8.2
PyPDF2==3.0.1
python-dotenv==1.0.1
tabula_py==2.9.3
uvicorn==0.22.0
psycopg2==2.9.9
sentence-transformers